Move files to where they are used
