--- RemoteSpiNNaker/RemoteSpiNNakerModel/src/main/java/uk/ac/manchester/cs/spinnaker/job/JobParameters.java
@@ -1,25 +1,0 @@
-/*
- * Copyright (c) 2014-2019 The University of Manchester
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License
- * along with this program.  If not, see <http://www.gnu.org/licenses/>.
- */
-package uk.ac.manchester.cs.spinnaker.job;
-
-/**
- * A type of Job to be executed. Note that the implementation must be annotated
- * with {@link JobParametersTypeName} if it is to be serialized or deserialized.
- */
-public interface JobParameters {
-    // No methods defined
-}
